Remove commented out code, add actual comments, add a diagnostic option

+++ WebResources/Headless/WebResources/headless.js
@@ -5,6 +5,14 @@ class Headless {
     constructor(bridge) {
         this._bridge = bridge;
 
+        // Diagnostic options.
+        //
+        // Don't delete Go Rest user after creation. This will cause the next
+        // user creation to fail due to duplicate email address. That in turn
+        // triggers the deletion of all matching users.
+        this._dontDelete = false;
+
+        // First tranche of code is only here to build the diagnostic UI.
         this._token = this._add_input(
             'Token', "Paste in your go-rest Access Token");
 
@@ -13,13 +21,19 @@ class Headless {
         this._add_command_button('SWAPI', {
             api:'star-wars', path:['people', 3]
         });
+
         this._add_command_button('go-rest GET', {
             api:'go-rest', path:['users', 20]
         });
         this._add_command_button('go-rest 401', {
             method:"POST", api:'go-rest', path:['users']
         });
-        // Query parameter isn't supported any more.
+        // Query parameter isn't supported by Go Rest.
+        //
+        // >   https://gorest.co.in/ says this:  
+        // >   This API supports only HTTP Bearer Tokens
+        //
+        // Code is left here in case it's useful for another API later.
         //
         // this._add_command_button('go-rest query parameter', {
         //     method:"POST", api:'go-rest', path:['users'],
@@ -29,15 +43,21 @@ class Headless {
             method:"POST", api:'go-rest', path:['users'], 'basic-auth': 'Bearer'
         });
 
+        this._hr = document.createElement('hr');
+        document.body.append(this._hr);
+        // End of main tranche of UI.
+
+        // Entry point for commands received from the native layer.
         this._bridge.receiveObjectCallback = command => Object.assign(
             this._command(command), {"confirm": "Main"});
 
-        this._hr = document.createElement('hr');
-        document.body.append(this._hr);
-
+        // Load secrets here, and for the native layer.
         import("./secrets.js")
         .then(secrets => {
+            // Copy the token into the input field.
             this._token.value = secrets.default.token;
+
+            // Send the token to the native layer.
             this._bridge.sendObject({"token": this._token.value});
         })
         .catch(error => {
@@ -46,10 +66,14 @@ class Headless {
             this._bridge.sendObject({"noToken": message});
         });
 
+        // Next line is also for the diagnostic UI though.
         document.getElementById('loading').firstChild.textContent = "Headless";
+
+        // Send the ready notification to the native layer.
         this._bridge.sendObject({"command": "ready"})
         .then(response => this._transcribe(response));
 
+        // Next line is also for the diagnostic UI though.
         this._token.focus();
     }
 
@@ -81,6 +105,12 @@ class Headless {
         return button;
     }
 
+    _transcribe(result) {
+        const pre = document.createElement('pre');
+        pre.append(JSON.stringify(result, undefined, 4));
+        this._hr.insertAdjacentElement('afterend', pre);
+    }
+
     _command(command) {
         const request = this._build_request(command);
         command.url = request.url;
@@ -97,9 +127,13 @@ class Headless {
         return command;
     }
 
+    // Create a standard Request object from a command dictionary.
     _build_request(command) {
+        // Convenience variable for the corresponding dictionary entry for the
+        // API to which the command relates.
         const base = Headless.apis[command.api];
 
+        // Assemble the HTTP header list.
         const headers = new Headers(base.headers);
         [command.headers].filter(
             headerObject => headerObject !== undefined
@@ -115,6 +149,13 @@ class Headless {
             );
         }
 
+        // Assemble the URI.
+        //
+        // It seems that you can't create a URL from nothing. This code starts
+        // with the document URL, which will usually be kind of a bogus value
+        // pointing to https:localhost, on Android, or local:, on iOS.
+        //
+        // Every part of the URL then gets replaced, so it should be OK.
         const uri = new URL(document.URL);
         uri.protocol = base.protocol;
         uri.hostname = base.hostname;
@@ -130,18 +171,22 @@ class Headless {
             );
         }
 
-        const request = new Request(
+        // Finally, instantiate and return a Request object.
+        return new Request(
             uri, {method: command.method, cache: 'no-cache', headers:headers});
-
-        return request;
     }
 
+    // Generate a JSON-able object from a Request instance.
     _request_JSON(request) {
         return {
             url: request.url, headers: [...request.headers.entries()]
         };
     }
 
+    // Main entry point for HTTP requests.
+    //
+    // This method takes different assumed actions depending on the request
+    // method and destination.
     async _request(request) {
         let errorMessage;
 
@@ -181,6 +226,10 @@ class Headless {
         return {error:errorMessage, request:this._request_JSON(request)};
     }
 
+    // Convenience method to send a request and get the JSON body.
+    //
+    // This method doesn't throw. Instead, it returns an object with an `error`
+    // property.
     async _fetch_JSON(request) {
         let response;
         try {
@@ -217,22 +266,17 @@ class Headless {
                 request:this._request_JSON(request),
             };
         }
-
     }
 
-    // _fetch_one(command, body) {
-    //     const fetchOptions = {
-    //         method:command.method,
-    //         headers:command.headers,
-    //         cache:"no-cache"
-    //     };
-    //     if (body !== undefined) {
-    //         fetchOptions.body = body;
-    //     }
-    //     return fetch(command.uri, fetchOptions)
-    //     .then(response => Promise.all([response, response.json()]));
-    // }
-
+    // Send an HTTP POST request, with Fetch, to create a user in Go Rest.
+    //
+    // -   If the request succeeds, remove the user from Go Rest, by sending an
+    //     HTTP DELETE request.
+    // -   If the request fails because there is already a user with the same
+    //     email address, delete that user and try again.
+    // -   If the request fails for any other reason, return a failure.
+    //
+    // This method has the same return style as the `_fetch_JSON` method.
     async _fetch_post_delete(request) {
         const userDetails = {
             name: 'Andy Warhol',
@@ -241,35 +285,27 @@ class Headless {
             status: 'Inactive'
         };
 
-
-        // return this._fetch_one(this._build_command({
-        //     method:"GET", api:'go-rest', path:['users']
-        // }))
-        // .then(([response, body]) => {
-        //     console.log('internal fetch', response, body);
-        //     let matchingEmail = [];
-        //     body.data.forEach(user => {
-        //         console.log(user.email);
-        //         if (user.email === userDetails.email) {
-        //             matchingEmail.push(user.id);
-        //         }
-        //     });
-        //     console.log('Matching email user identifiers:', matchingEmail);
-
-
-
         const userBlob = new Blob(
             [JSON.stringify(userDetails, null, 2)],
             {type : 'application/json'}
         );
-        console.log('posting', userDetails);
+        // console.log('posting', userDetails);
 
         const firstRequest = new Request(request, {body: userBlob});
         let response = await this._fetch_JSON(firstRequest);
-        console.log(response);
+        // console.log(response);
 
         if (response.body.code === 422) {
-            await this._delete_matches(request, userDetails);
+            // 422 is returned for anything invalid, including duplicate email
+            // address. For now, assume a duplicate email address is the problem
+            // and delete any matches.
+            const matchResponse = await this._delete_matches(
+                request, userDetails);
+            if (matchResponse.error !== undefined) {
+                return matchResponse;
+            }
+
+            // Try again.
             const secondRequest = new Request(request, {body: userBlob});
             response = await this._fetch_JSON(secondRequest);
         }
@@ -278,51 +314,33 @@ class Headless {
             return {error: response, request: this._request_JSON(firstRequest)};
         }
 
+        if (this._dontDelete) {
+            return response;
+        }
+
         // User entry was created OK. Delete it so that the email address isn't
         // in use and the create can be run again.
         const deleteResponse = await this._delete_id(
             request, response.body.data.id);
         if (deleteResponse.error === undefined) {
+            // Delete is OK, so everything is OK. Note the deletion and return
+            // the response to the POST command.
             response.deleted = "after";
             return response;
         }
+
+        // Delete failed. Return the failure response.
         return deleteResponse;
-
-        // return Promise.resolve({error:"Not implemented."});
-        // fetch(
-        //     [Headless.apis["go-rest"].base, "users"].join('/'),
-        //     {method:"POST", body: blob, headers:command.headers}
-        // )
-        // .then(response => Promise.all([response, response.json()]))
-        // .then(([response, body]) => {
-        //     console.log(response, body)
-        // })
-        // .catch(error => {
-        //     console.log(error);
-        // });
-
-
-
-        // console.log('in time out for send.');
-        // .then(response => Promise.all([response, response.json()]))
-        // .then(([response, body]) => {
-        //     // console.log(response, body)
-        //     const result = {
-        //         ok:response.ok, body:body, status:response.status,
-        //         statusText:response.statusText
-        //     };
-        //     this._transcribe(result);
-        //     this._bridge.sendObject(result);
-        // })
-        // .catch(error => {
-        //     // console.log(error);
-        //     const result = {error:`${error}`};
-        //     this._transcribe(result);
-        //     this._bridge.sendObject(result);
-        // });
-
     }
 
+    // Delete every Go Rest user with a specified email address.
+    //
+    // This method uses the paging feature of the Go Rest API. Go Rest doesn't
+    // seem to support querying for email address, so the code here retrieves
+    // all the users, page by page. When it finds a matching user, it sends a
+    // delete command and then starts again.
+    //
+    // This method has the same return style as the `_fetch_JSON` method.
     async _delete_matches(
         request, userDetails, duplicates=undefined, startPage=1
     ) {
@@ -342,21 +360,16 @@ class Headless {
         const getRequest = new Request(pageURI, {
             headers: request.headers, cache: request.cache
         });
-        console.log(this._request_JSON(getRequest));
+        // console.log(this._request_JSON(getRequest));
         
         const pageResponse = await this._fetch_JSON(getRequest);
         if (pageResponse.error !== undefined) {
             return pageResponse;
         }
-        
-        // (this._build_command({
-        //     method:"GET", api:'go-rest', path:['users'],
-        //     parameters:{page:startPage}
-        // }));
-        console.log('match page', startPage, pageResponse.body);
+        // console.log('match page', startPage, pageResponse.body);
 
         if (startPage > pageResponse.body.meta.pagination.pages) {
-            return duplicates;
+            return {deleted:duplicates};
         }
 
         let deleteNow = null;
@@ -367,24 +380,31 @@ class Headless {
                 break;
             }
         }
+        this._transcribe({
+            deleting:userDetails.email,
+            page:startPage,
+            found:deleteNow !== null 
+        });
 
         if (deleteNow !== null) {
-            console.log('Deleting', startPage, deleteNow);
+            // console.log('Deleting', startPage, deleteNow);
             const deleteResponse = await this._delete_id(request, deleteNow);
-            console.log(deleteResponse);
-
-            // console.log('Deleted', deleteResult)
-        // Delete the one to be deleted now, then return delete_matches starting at page 1.
-
-
+            // console.log(deleteResponse);
+            if (deleteResponse.error !== undefined) {
+                return deleteResponse;
+            }
+    
             duplicates.push(deleteNow);
-            // startPage = 0;
+            startPage = 0;
         }
 
         return this._delete_matches(
             request, userDetails, duplicates, startPage + 1);
     }
 
+    // Delete the specified Go Rest user.
+    //
+    // This method has the same return style as the `_fetch_JSON` method.
     async _delete_id(request, userIdentifier) {
         const deleteURI = new URL(request.url);
         deleteURI.pathname = [
@@ -394,20 +414,11 @@ class Headless {
             headers: request.headers, cache: request.cache, method:"DELETE"
         });
         return await this._fetch_JSON(deleteRequest);
-    // return this._fetch_one(this._build_command({
-        //     method:"DELETE", api:'go-rest', path:['users', `${userIdentifier}`]
-        // }));
-    }
-
-    _transcribe(result) {
-        const pre = document.createElement('pre');
-        pre.append(JSON.stringify(result, undefined, 4));
-        this._hr.insertAdjacentElement('afterend', pre);
     }
 }
 Headless.apis = {
     'star-wars': {
-        protocol: "http",
+        protocol: "https",
         hostname: "swapi.dev",
         path: ["api"],
         headers: {}
